Phase 2: Core integration — prepared statement caching in executor
Implement the core prepared statement caching path in the adaptive executor:

1. Query template preservation (citus_clauses.c/h):
   - Add EVALUATE_FUNCTIONS mode to CoordinatorEvaluationMode enum
   - New ExecuteCoordinatorEvaluableFunctions() evaluates functions
     while preserving Param nodes for parameterized queries

2. Task metadata (multi_physical_planner.h, citus_custom_scan.c):
   - Add preparedStatementPlanId and jobQueryForPrepare fields to Task
   - CitusBeginReadOnlyScan and CitusBeginModifyScan save pre-evaluation
     job query and annotate tasks with plan ID for cache lookup
   - INSERT commands excluded (query trees have special RTEs that
     pg_get_query_def cannot deparse)

3. SendNextQuery prepared statement path (adaptive_executor.c):
   - New if-block at top of SendNextQuery() checks
     EnablePreparedStatementCaching && task->jobQueryForPrepare
   - Cache miss: UpdateRelationToShardNames + pg_get_query_def to
     construct parameterized shard query, PQprepare via SendRemotePrepare,
     cache the entry, then PQsendQueryPrepared via SendRemotePreparedQuery
   - Cache hit: reuse cached statement name with PQsendQueryPrepared
   - Fallback: goto plain_sql when cache is full
   - Skip MarkUnreferencedExternParams when caching enabled to preserve
     original parameter types

4. libpq wrappers (remote_commands.c/h):
   - SendRemotePrepare: synchronous PQprepare wrapper
   - SendRemotePreparedQuery: async PQsendQueryPrepared wrapper

5. Cache key fix (prepared_statement_cache.c/h):
   - Changed planId from uint32 to uint64 to match DistributedPlan.planId

+++ src/backend/distributed/executor/prepared_statement_cache.c
@@ -57,7 +57,7 @@ PreparedStatementCacheCreate(void)
  * Returns the entry if found, NULL otherwise.
  */
 PreparedStatementCacheEntry *
-PreparedStatementCacheLookup(HTAB *cache, uint32 planId, uint64 shardId)
+PreparedStatementCacheLookup(HTAB *cache, uint64 planId, uint64 shardId)
 {
 	PreparedStatementCacheKey key;
 
@@ -82,7 +82,7 @@ PreparedStatementCacheLookup(HTAB *cache, uint32 planId, uint64 shardId)
  * (stmtName, paramTypes, paramCount, parameterizedQueryString).
  */
 PreparedStatementCacheEntry *
-PreparedStatementCacheInsert(HTAB *cache, uint32 planId, uint64 shardId)
+PreparedStatementCacheInsert(HTAB *cache, uint64 planId, uint64 shardId)
 {
 	if (hash_get_num_entries(cache) >= MAX_CACHED_STMTS_PER_CONNECTION)
 	{
